Time: 713 ms (5%), Space: 44.9 MB (6.35%) - LeetHub

+++ 0713-subarray-product-less-than-k/0713-subarray-product-less-than-k.cpp
@@ -3,7 +3,7 @@ public:
     int numSubarrayProductLessThanK(vector<int>& nums, int k) {
         int count =0;
         for(int i=0; i<nums.size(); i++){
-            long long curr_mul =1;
+            long curr_mul =1;
             for(int j=i; j<nums.size(); j++){
                 curr_mul *= nums[j];
                 if(curr_mul >=k){
